+++ R/process-methods.R
@@ -39,7 +39,7 @@ process.sequential_chat <- function(chat_env,
 #' @param file Path to save state file
 #' @param workers Number of parallel workers (default upper limit is `parallel::detectCores()`)
 #' @param chunk_size Number of prompts each worker processes at a time (default: 10)
-#' @param max_chunk_attempts Maximum retries per failed chunk
+#' @param max_chunk_tries Maximum tries per failed chunk
 #' @param progress Whether to show progress bars
 #' @param beep Whether to play a sound on completion
 #' @param echo Whether to display chat outputs (when `progress` is `FALSE`)
@@ -53,7 +53,7 @@ process.future_chat <- function(chat_env,
                                 file = tempfile("chat_", fileext = ".rds"),
                                 workers = NULL,
                                 chunk_size = 10,
-                                max_chunk_attempts = 3L,
+                                max_chunk_tries = 2L,
                                 beep = TRUE,
                                 progress = TRUE,
                                 echo = FALSE,
@@ -63,7 +63,7 @@ process.future_chat <- function(chat_env,
   } else {
     chat_obj <- chat_env
   }
-  
+
   process_future(
     chat_obj = chat_obj,
     prompts = prompts,
@@ -71,7 +71,7 @@ process.future_chat <- function(chat_env,
     file = file,
     workers = workers,
     chunk_size = chunk_size,
-    max_chunk_attempts = max_chunk_attempts,
+    max_chunk_tries = max_chunk_tries,
     beep = beep,
     progress = progress,
     echo = echo,